requirements.txt
aiofiles==25.1.0
aiogram==3.27.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
beautifulsoup4==4.14.3
certifi==2026.4.22
cffi==2.0.0
chardet==3.0.4
charset-normalizer==3.4.7
cryptography==47.0.0
deep-translator==1.11.4
feedparser==6.0.12
filetype==1.2.0
frozenlist==1.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.195.0
google-auth==2.50.0
google-auth-httplib2==0.3.1
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
greenlet==3.5.0
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
hyperframe==5.2.0
idna==2.10
Js2Py_3.13==0.74.1
lxml==6.1.0
m3u8==6.0.0
magic-filter==1.0.12
multidict==6.7.1
ntscraper==0.4.0
propcache==0.4.1
proto-plus==1.27.2
protobuf==5.29.6
pyaes==1.6.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyjsparser==2.7.1
pyotp==2.9.0
pyparsing==3.3.2
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.2.2
requests==2.33.1
rfc3986==1.5.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8.3
SQLAlchemy==2.0.49
TgCrypto==1.2.5
tqdm==4.67.3
twikit==2.3.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
webvtt-py==0.5.1
yarl==1.23.0
